feat: add classes to status
