+++ scripts/utils.js
@@ -31,7 +31,23 @@ function walk(dir, callback) {
   });
 }
 
+function slugify(str) {
+  str = str.replace(/^\s+|\s+$/g, ''); // trim leading/trailing white space
+  str = str.toLowerCase(); // convert string to lowercase
+  str = str.replace(/[^a-z0-9 -]/g, '') // remove any non-alphanumeric characters
+           .replace(/\s+/g, '-') // replace spaces with hyphens
+           .replace(/-+/g, '-'); // remove consecutive hyphens
+  return str;
+}
+
+const toTitleCase = (string) => {
+  string = string.toLowerCase()
+  return string.charAt(0).toUpperCase() + string.slice(1);
+};
+
 module.exports = {
   searchInDir,
   walk,
+  slugify,
+  toTitleCase
 };
